refactor board dnd logic into hook
